Remove CodeHashing capabilities from memo's hashing.py
